refactor: move Head and LoadingOrError component into Common
[why] to group common component together
